Manage parameters and options in jenkinsfile
Not all parameters were moved from config.xml yet.

But we moved buildDiscarder to jenkinsfile and removed them from
config.xml. We also set a more conservative setting of 7 days or 10
builds to preserve disk space.

We added two new groovy scripts to set the description and verify if
parameter has changed so that we abort the build in case the parameter
list were changed and the user must retrigger the job with the new
parameters.

Signed-off-by: Qais Yousef <qyousef@layalina.io>